rename quality and tool use composite evals
