+++ assets/evaluators/builtin/output_quality/__init__.py
@@ -1,4 +1,4 @@
 # Copyright (c) Microsoft Corporation.
 # Licensed under the MIT License.
 
-"""Quality Composite Evaluator package."""
+"""Output Quality Evaluator package."""
