+++ assets/evaluators/builtin/output_quality/evaluator/__init__.py
@@ -1,4 +1,4 @@
 # Copyright (c) Microsoft Corporation.
 # Licensed under the MIT License.
 
-"""Quality Composite Evaluator implementation package."""
+"""Output Quality Evaluator implementation package."""
