+++ assets/evaluators/builtin/output_quality/evaluator/output_quality_multi_turn.prompty
@@ -1,5 +1,5 @@
 ---
-name: Quality Composite Evaluator (Multi-Turn)
+name: Output Quality Evaluator (Multi-Turn)
 description: >
   One judge call scoring SIX independent quality rubrics (fluency, coherence, intent_resolution, task_adherence, groundedness, task_completion) over a full multi-turn user-agent conversation, and identifying the earliest failed user turn per rubric.
 model:
