+++ assets/evaluators/builtin/output_quality/spec.yaml
@@ -1,17 +1,17 @@
 type: "evaluator"
-id: "azureml://registries/azureml/evaluators/builtin.tool_use_composite/versions/1"
-name: "builtin.tool_use_composite"
+id: "azureml://registries/azureml/evaluators/builtin.output_quality/versions/1"
+name: "builtin.output_quality"
 version: 1
-displayName: "Tool-Use-Composite-Evaluator-(Preview)"
-description: "A composite evaluator that batches the tool_call_accuracy, tool_call_success, tool_input_accuracy, tool_output_utilization, and tool_selection evaluators into a single LLM call. The primary tool_use_composite score fails when any evaluated evaluator fails, while the evaluator keys preserve the LLM results and derived pass/fail fields. Supports both single-turn (query/response) and multi-turn (messages) agent interactions."
+displayName: "Output-Quality-Evaluator-(Preview)"
+description: "A composite evaluator that batches fluency, coherence, intent_resolution, task_adherence, groundedness, and task_completion into a single LLM call. The primary output_quality score fails when any evaluated evaluator fails, while the evaluator keys preserve the LLM results and derived pass/fail fields. Supports single-turn query/response and multi-turn messages inputs."
 evaluatorType: "builtin"
 evaluatorSubType: "code"
-categories: ["agents"]
+categories: ["quality", "agents"]
+supportedEvaluationLevels: ["conversation", "turn"]
 tags:
   provider: "Microsoft"
   is_continuous_scenario: "true"
-  composite_evaluator: "tool_call_accuracy, tool_call_success, tool_input_accuracy, tool_output_utilization, tool_selection"
-supportedEvaluationLevels: ["conversation", "turn"]
+  composite_evaluator: "fluency, coherence, intent_resolution, task_adherence, groundedness, task_completion"
 initParameterSchema:
   type: "object"
   properties:
@@ -19,8 +19,8 @@ initParameterSchema:
       type: "string"
     threshold:
       type: "number"
-      minimum: 0
-      maximum: 1
+      minimum: 1
+      maximum: 5
       multipleOf: 1
     evaluation_level:
       type: "string"
@@ -40,47 +40,52 @@ dataMappingSchema:
         - type: "array"
           items:
             type: "object"
-    messages:
-      type: "array"
-      items:
-        type: "object"
     tool_definitions:
       anyOf:
         - type: "object"
         - type: "array"
           items:
             type: "object"
+    messages:
+      type: "array"
+      items:
+        type: "object"
   anyOf:
-    - required: ["query", "response", "tool_definitions"]
-    - required: ["messages", "tool_definitions"]
+    - required: ["query", "response"]
+    - required: ["messages"]
 outputSchema:
-  tool_use_composite:
+  output_quality:
     type: "boolean"
     desirable_direction: "increase"
     min_value: 0
     max_value: 1
     is_primary: true
-  tool_call_accuracy:
+  fluency:
     type: "ordinal"
     desirable_direction: "increase"
     min_value: 1
     max_value: 5
-  tool_call_success:
+  coherence:
+    type: "ordinal"
+    desirable_direction: "increase"
+    min_value: 1
+    max_value: 5
+  intent_resolution:
+    type: "ordinal"
+    desirable_direction: "increase"
+    min_value: 1
+    max_value: 5
+  task_adherence:
     type: "boolean"
     desirable_direction: "increase"
     min_value: 0
     max_value: 1
-  tool_input_accuracy:
-    type: "boolean"
+  groundedness:
+    type: "ordinal"
     desirable_direction: "increase"
-    min_value: 0
-    max_value: 1
-  tool_output_utilization:
-    type: "boolean"
-    desirable_direction: "increase"
-    min_value: 0
-    max_value: 1
-  tool_selection:
+    min_value: 1
+    max_value: 5
+  task_completion:
     type: "boolean"
     desirable_direction: "increase"
     min_value: 0
