+++ assets/evaluators/builtin/tool_use_quality/asset.yaml
@@ -6,4 +6,4 @@ test:
   pytest:
     enabled: true
     conda_environment: ../../tests/conda.yaml
-    tests_dir: ../../tests/test_evaluators_behavior/test_tool_use_composite_evaluator_behavior.py
+    tests_dir: ../../tests/test_evaluators_behavior/test_tool_use_quality_evaluator_behavior.py
