+++ assets/evaluators/builtin/tool_use_quality/evaluator/__init__.py
@@ -1,4 +1,4 @@
 # Copyright (c) Microsoft Corporation.
 # Licensed under the MIT License.
 
-"""Tool Use Composite Evaluator package."""
+"""Tool Use Quality Evaluator implementation package."""
