+++ assets/evaluators/builtin/tool_use_quality/evaluator/tool_use_quality.prompty
@@ -1,5 +1,5 @@
 ---
-name: Tool Use Composite Evaluator
+name: Tool Use Quality Evaluator
 description: >
   A single composite evaluator that scores ALL FIVE prod tool evaluators
   (tool_call_accuracy, tool_call_success, tool_input_accuracy,
