+++ assets/evaluators/builtin/tool_use_quality/evaluator/tool_use_quality_multi_turn.prompty
@@ -1,5 +1,5 @@
 ---
-name: Tool Use Composite Evaluator (Multi-Turn)
+name: Tool Use Quality Evaluator (Multi-Turn)
 description: >
   Multi-turn variant of the tool use composite evaluator. Scores ALL FIVE
   prod tool evaluators (tool_call_accuracy, tool_call_success,
