+++ assets/evaluators/builtin/tool_use_quality/spec.yaml
@@ -1,17 +1,17 @@
 type: "evaluator"
-id: "azureml://registries/azureml/evaluators/builtin.quality_composite/versions/1"
-name: "builtin.quality_composite"
+id: "azureml://registries/azureml/evaluators/builtin.tool_use_quality/versions/1"
+name: "builtin.tool_use_quality"
 version: 1
-displayName: "Quality-Composite-Evaluator-(Preview)"
-description: "A composite evaluator that batches fluency, coherence, intent_resolution, task_adherence, groundedness, and task_completion into a single LLM call. The primary quality_composite score fails when any evaluated evaluator fails, while the evaluator keys preserve the LLM results and derived pass/fail fields. Supports single-turn query/response and multi-turn messages inputs."
+displayName: "Tool-Use-Quality-Evaluator-(Preview)"
+description: "A composite evaluator that batches the tool_call_accuracy, tool_call_success, tool_input_accuracy, tool_output_utilization, and tool_selection evaluators into a single LLM call. The primary tool_use_quality score fails when any evaluated evaluator fails, while the evaluator keys preserve the LLM results and derived pass/fail fields. Supports both single-turn (query/response) and multi-turn (messages) agent interactions."
 evaluatorType: "builtin"
 evaluatorSubType: "code"
-categories: ["quality", "agents"]
-supportedEvaluationLevels: ["conversation", "turn"]
+categories: ["agents"]
 tags:
   provider: "Microsoft"
   is_continuous_scenario: "true"
-  composite_evaluator: "fluency, coherence, intent_resolution, task_adherence, groundedness, task_completion"
+  composite_evaluator: "tool_call_accuracy, tool_call_success, tool_input_accuracy, tool_output_utilization, tool_selection"
+supportedEvaluationLevels: ["conversation", "turn"]
 initParameterSchema:
   type: "object"
   properties:
@@ -19,8 +19,8 @@ initParameterSchema:
       type: "string"
     threshold:
       type: "number"
-      minimum: 1
-      maximum: 5
+      minimum: 0
+      maximum: 1
       multipleOf: 1
     evaluation_level:
       type: "string"
@@ -40,52 +40,47 @@ dataMappingSchema:
         - type: "array"
           items:
             type: "object"
+    messages:
+      type: "array"
+      items:
+        type: "object"
     tool_definitions:
       anyOf:
         - type: "object"
         - type: "array"
           items:
             type: "object"
-    messages:
-      type: "array"
-      items:
-        type: "object"
   anyOf:
-    - required: ["query", "response"]
-    - required: ["messages"]
+    - required: ["query", "response", "tool_definitions"]
+    - required: ["messages", "tool_definitions"]
 outputSchema:
-  quality_composite:
+  tool_use_quality:
     type: "boolean"
     desirable_direction: "increase"
     min_value: 0
     max_value: 1
     is_primary: true
-  fluency:
+  tool_call_accuracy:
     type: "ordinal"
     desirable_direction: "increase"
     min_value: 1
     max_value: 5
-  coherence:
-    type: "ordinal"
-    desirable_direction: "increase"
-    min_value: 1
-    max_value: 5
-  intent_resolution:
-    type: "ordinal"
-    desirable_direction: "increase"
-    min_value: 1
-    max_value: 5
-  task_adherence:
+  tool_call_success:
     type: "boolean"
     desirable_direction: "increase"
     min_value: 0
     max_value: 1
-  groundedness:
-    type: "ordinal"
+  tool_input_accuracy:
+    type: "boolean"
     desirable_direction: "increase"
-    min_value: 1
-    max_value: 5
-  task_completion:
+    min_value: 0
+    max_value: 1
+  tool_output_utilization:
+    type: "boolean"
+    desirable_direction: "increase"
+    min_value: 0
+    max_value: 1
+  tool_selection:
     type: "boolean"
     desirable_direction: "increase"
     min_value: 0
