+++ assets/evaluators/tests/test_evaluators_quality/test_tool_use_quality_evaluator_quality.py
@@ -3,7 +3,7 @@
 
 """Quality tests for the Tool Use Evaluation Suite (single-turn) with real flow execution.
 
-ToolUseCompositeEvaluator batches five evaluators into a single LLM call, so its output shape
+ToolUseQualityEvaluator batches five evaluators into a single LLM call, so its output shape
 (one primary score plus five raw evaluator result objects) differs from the single-evaluator
 shape assumed by ``BaseQualityEvaluatorRunner``. This file therefore builds the evaluator
 directly (same env-var + ``DefaultAzureCredential`` pattern used by
@@ -21,32 +21,32 @@ import pytest
 from azure.ai.evaluation import AzureOpenAIModelConfiguration
 from azure.identity import DefaultAzureCredential
 
-from ...builtin.tool_use_composite.evaluator._tool_use_composite import ToolUseCompositeEvaluator
+from ...builtin.tool_use_quality.evaluator._tool_use_quality import ToolUseQualityEvaluator
 
 
-def _make_real_evaluator() -> ToolUseCompositeEvaluator:
+def _make_real_evaluator() -> ToolUseQualityEvaluator:
     model_config = AzureOpenAIModelConfiguration(
         azure_endpoint=os.getenv("AZURE_OPENAI_ENDPOINT"),
         azure_deployment=os.getenv("AZURE_OPENAI_DEPLOYMENT"),
         api_version=os.getenv("AZURE_OPENAI_API_VERSION", "2024-08-01-preview"),
     )
-    return ToolUseCompositeEvaluator(model_config=model_config, credential=DefaultAzureCredential())
+    return ToolUseQualityEvaluator(model_config=model_config, credential=DefaultAzureCredential())
 
 
 def _assert_evaluator_passed(result, name: str):
-    evaluator = result["tool_use_composite_evaluators"][name]
+    evaluator = result["tool_use_quality_evaluators"][name]
     assert evaluator["status"] == "completed", f"{name} should be completed, got {evaluator.get('status')}"
     assert evaluator["score"] is not None, f"{name} expected a score, reason: {evaluator.get('reason')}"
 
 
 def _assert_evaluator_failed(result, name: str):
-    evaluator = result["tool_use_composite_evaluators"][name]
+    evaluator = result["tool_use_quality_evaluators"][name]
     assert evaluator["status"] == "completed", f"{name} should be completed, got {evaluator.get('status')}"
     assert evaluator["score"] is not None, f"{name} expected a score, reason: {evaluator.get('reason')}"
 
 
 def _assert_evaluator_skipped(result, name: str):
-    evaluator = result["tool_use_composite_evaluators"][name]
+    evaluator = result["tool_use_quality_evaluators"][name]
     assert evaluator["status"] == "skipped", f"{name} expected skipped, got {evaluator.get('status')}"
     assert evaluator["score"] is None
 
@@ -69,8 +69,8 @@ WEATHER_QUERY = [
 
 
 @pytest.mark.quality
-class TestToolUseCompositeEvaluatorQuality:
-    """Quality tests for ToolUseCompositeEvaluator with single-turn (query/response) input."""
+class TestToolUseQualityEvaluatorQuality:
+    """Quality tests for ToolUseQualityEvaluator with single-turn (query/response) input."""
 
     def test_all_evaluators_pass_for_correct_single_tool_call(self):
         """A single correct, successful, well-utilized, well-selected tool call passes all five evaluators."""
@@ -109,8 +109,8 @@ class TestToolUseCompositeEvaluatorQuality:
             "tool_selection",
         ):
             _assert_evaluator_passed(result, name)
-        assert result["tool_use_composite"] == 1
-        assert result["tool_use_composite_evaluators"]["tool_call_accuracy"]["score"] == 5
+        assert result["tool_use_quality"] == 1
+        assert result["tool_use_quality_evaluators"]["tool_call_accuracy"]["score"] == 5
 
     def test_tool_input_accuracy_fails_on_fabricated_parameter(self):
         """A tool call with a fabricated parameter (not present in the conversation) fails tool_input_accuracy."""
@@ -142,7 +142,7 @@ class TestToolUseCompositeEvaluatorQuality:
             tool_definitions=WEATHER_TOOL_DEFINITIONS,
         )
         _assert_evaluator_failed(result, "tool_input_accuracy")
-        assert result["tool_use_composite"] == 0
+        assert result["tool_use_quality"] == 0
 
     def test_tool_call_success_fails_on_tool_error(self):
         """A tool call whose result indicates a technical error fails tool_call_success."""
@@ -174,7 +174,7 @@ class TestToolUseCompositeEvaluatorQuality:
             tool_definitions=WEATHER_TOOL_DEFINITIONS,
         )
         _assert_evaluator_failed(result, "tool_call_success")
-        assert result["tool_use_composite"] == 0
+        assert result["tool_use_quality"] == 0
 
     def test_tool_selection_fails_when_wrong_tool_used(self):
         """Selecting a tool irrelevant to the user's request fails tool_selection."""
@@ -217,7 +217,7 @@ class TestToolUseCompositeEvaluatorQuality:
             tool_definitions=tool_definitions,
         )
         _assert_evaluator_failed(result, "tool_selection")
-        assert result["tool_use_composite"] == 0
+        assert result["tool_use_quality"] == 0
 
     def test_all_evaluators_skipped_when_no_tool_calls_made(self):
         """No tool calls in the response yields skipped/not_applicable results for tool-dependent evaluators."""
